Split compliance checks into focused modules
