+++ rclip/compliance/__init__.py
@@ -1,0 +1,1 @@
+"""Build-time licence and distribution compliance support."""
